// ui/admin/src/models/filterSnapshot.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Wanaku API
 * Wanaku MCP proxy management API
 * OpenAPI spec version: 0.3.0
 */
import type { DurationSnapshot } from "./durationSnapshot";

export interface FilterSnapshot {
  duration: DurationSnapshot;
  /** @minimum 0 */
  errors: number;
  /** @minimum 0 */
  requests_continue: number;
  /** @minimum 0 */
  requests_other: number;
  /** @minimum 0 */
  requests_reject: number;
}